HtmlFragmentSelection: Move to util package
